+++ protos/embed_test.go
@@ -44,7 +44,7 @@ func Test_DataNodeBindings(t *testing.T) {
 	t.Run("CoreBindings should return the core http bindings", func(t *testing.T) {
 		bindings, err := protos.DataNodeBindings()
 		require.NoError(t, err)
-		wantCount := 83
+		wantCount := 81
 
 		assert.Len(t, bindings.HTTP.Rules, wantCount)